Some ui tweaks
The title on the error field doesn't need to be red for the error state.

Brackets and spaces need to be removed from the phone number input field. We can mask the input and put it in that format

The title for "Contact email", "Phone number" and "Support instructions" needs to be h4

Only use upper case for the first word in the titles