Looks like I shouldn't do both = false and ?
https://stackoverflow.com/questions/22680391/cannot-implicitly-convert-type-bool-to-bool-an-explicit-conversion-exists

bool? is not a bool. It is in reality a Nullable<bool> http://msdn.microsoft.com/en-us/library/b3h38hb0(v=vs.110).aspx

If you need the bool value then you should either cast like you are doing or call the .Value property on the bool?. There is also a .HasValue property you can check to make sure that it is not null.

If IsChecked is null, this line will error.