CS-2786 - Define and create custom context for ADSP forms
* Defined the interface to a custom context using React’s context feature. The context supports:

providers for registers of master data
providers for file upload/download

interface uses a urn as reference to the file for upload and download

User access token or cookie credential for accessing provider API is handled in the implementation of the providers.

Included the context in the jsonforms-component library as appropriate so that in the future teams can import the library and have everything they need to set up json forms with ADSP customizations and extensions in their own app

Configured form should work even if the context provider is not present.

application may use form without need for any of the extension features, so form should successfully render even if the context is not present.