feat(notification-service): support collection of criteria objects for multiple specific subscriptions
Currently the database schema includes a unique constraint on notification type and subscriber for subscriptions, so that each subscriber can only have a single subscription per type. This is not easily changed in a migration, and so to support a subscriber with subscription to multiple but specific forms use a collection of criteria instead. The subscription sends if any criteria object in the collection matches.