CS-2889 - Fix regular login
If the parameter 'kc_idp_hint' does not exist in the urlParams object, calling urlParams.get('kc_idp_hint') will return null. If you then pass null to encodeURIComponent, it will not cause an error, but the result will be the string "null" being encoded.

This results in the logic after than to incorrectly determine than a
parameter is there when none was passed, preventing the regular login

This code checks to ensure there is a parameter. If there is none, it
will set it idpFromUrl to null, and the core will execute as intended