feat(adsp-service-sdk): include qualified service client roles
Add qualified name for roles of the current service the SDK is initialized against in addition to unqualified name to support cases where a client role of the service is included in configuration. Configuration need to use qualified names for end user configuration experience; otherwise the roles can be ambiguous across different services.