feat(directory-service): resolve error handling
If resolve encounters a missing resource, then delete it for consistency. Note that the directory resource API does not check if an API resource exists, but the resolve job will delete it after for eventual consistency if an associate type exists.

If resolve fails completed (no retries), signal a domain event so the issue is visible to the tenant.