feat(pdf-service): support job context
Allow request to include context in job which is then used in pdf generation events. This is to support triggering on pdf generation events based on a context criteria. The recordId has more specific meaning since it is used to associate the output file to a record and not just for event based trigger.